feat: move PlayButton to Components
